% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezDiffusion.R
\name{ez_diffusion_model}
\alias{ez_diffusion_model}
\title{The EZ-diffusion model}
\usage{
ez_diffusion_model(data, id, rt, correct, within = NULL, s = 0.1)
}
\arguments{
\item{data}{A \code{data.frame} that contains the data.}

\item{id}{Character. The name of the variable that contains the participant identifier.}

\item{rt}{Character. The name of the variable that contains the reaction times.}

\item{correct}{Character. The name of the variable that codes correct vs. erroneous decision. Needs to be 0 = erroneous and 1 = correct.}

\item{within}{Character. Possibly a vector of names of variables that contain within-subjects factors.}

\item{s}{Numeric. An optional scaling constant.}
}
\description{
This is the convenient wrwapper function for the EZ-diffusion model
}
\examples{
data <- data.frame(
  id = rep(1:150, each = 100)
  , rt = runif(150 * 100, min = .2, max = .8)
  , correct = sample(0:1, size = 150 * 100, replace = TRUE)
)
ez_diffusion_model(data = data, id = "id", rt = "rt", correct = "correct")
}
