% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezDiffusion.R
\name{ezDiffusion}
\alias{ezDiffusion}
\title{EZ Diffusion}
\usage{
ezDiffusion(proportion_correct, rt_variance, rt_mean, s = 0.1, data)
}
\arguments{
\item{proportion_correct}{Character. The name of the variable that contains proportion of correct trials.}

\item{rt_variance}{Character. The name of the variable that contains the variance of reaction times for correct decisions}

\item{rt_mean}{Character. The name of the variable that contains the mean of reaction times for correct decisions.}

\item{s}{Numeric. A scaling constant. Take a look at the literature, because different
authors use different scaling constants.}

\item{data}{A \code{data.frame} that conitains the above-specified variables.}
}
\description{
Calculate the parameters of the EZ diffusion model from mrt, vrt, pc...
}
\examples{
data <- data.frame(
  prop_correct = runif(n = 100, min = 0.2, max = .8)
  , rt_mean = runif(n = 100, min = .2, max = .8)
  , rt_var = runif(n = 100, min = .01, max = .1)
)

ezDiffusion(
  data = data
  , proportion_correct = "prop_correct"
  , rt_variance = "rt_var"
  , rt_mean = "rt_mean"
 )

}
